Fix other tests

+++ UnitsNet.Tests/CustomCode/BitRateTests.cs
@@ -29,37 +29,37 @@ namespace UnitsNet.Tests.CustomCode
     {
         protected override bool SupportsSIUnitSystem => false;
 
-        protected override decimal BitsPerSecondInOneBitPerSecond => 1m;
-        protected override decimal BytesPerSecondInOneBitPerSecond => 1.25E-1m;
+        protected override double BitsPerSecondInOneBitPerSecond => 1;
+        protected override double BytesPerSecondInOneBitPerSecond => 1.25E-1;
 
-        protected override decimal KilobitsPerSecondInOneBitPerSecond => 1E-3m;
-        protected override decimal KilobytesPerSecondInOneBitPerSecond => 1.25E-4m;
-        protected override decimal KibibitsPerSecondInOneBitPerSecond => 0.0009765625m;
-        protected override decimal KibibytesPerSecondInOneBitPerSecond => 0.0001220703125m;
+        protected override double KilobitsPerSecondInOneBitPerSecond => 1E-3;
+        protected override double KilobytesPerSecondInOneBitPerSecond => 1.25E-4;
+        protected override double KibibitsPerSecondInOneBitPerSecond => 0.0009765625;
+        protected override double KibibytesPerSecondInOneBitPerSecond => 0.0001220703125;
 
-        protected override decimal MegabitsPerSecondInOneBitPerSecond => 1E-6m;
-        protected override decimal MegabytesPerSecondInOneBitPerSecond => 1.25E-07m;
-        protected override decimal MebibitsPerSecondInOneBitPerSecond => 9.5367431640625E-07m;
-        protected override decimal MebibytesPerSecondInOneBitPerSecond => 1.19209289550781E-07m;
+        protected override double MegabitsPerSecondInOneBitPerSecond => 1E-6;
+        protected override double MegabytesPerSecondInOneBitPerSecond => 1.25E-07;
+        protected override double MebibitsPerSecondInOneBitPerSecond => 9.5367431640625E-07;
+        protected override double MebibytesPerSecondInOneBitPerSecond => 1.19209289550781E-07;
 
-        protected override decimal GigabitsPerSecondInOneBitPerSecond => 1E-9m;
-        protected override decimal GigabytesPerSecondInOneBitPerSecond => 1.25E-10m;
-        protected override decimal GibibitsPerSecondInOneBitPerSecond => 9.31322574615479E-10m;
-        protected override decimal GibibytesPerSecondInOneBitPerSecond => 1.16415321826935E-10m;
+        protected override double GigabitsPerSecondInOneBitPerSecond => 1E-9;
+        protected override double GigabytesPerSecondInOneBitPerSecond => 1.25E-10;
+        protected override double GibibitsPerSecondInOneBitPerSecond => 9.31322574615479E-10;
+        protected override double GibibytesPerSecondInOneBitPerSecond => 1.16415321826935E-10;
 
-        protected override decimal TerabitsPerSecondInOneBitPerSecond => 1E-12m;
-        protected override decimal TerabytesPerSecondInOneBitPerSecond => 1.25E-13m;
-        protected override decimal TebibitsPerSecondInOneBitPerSecond => 9.09494701772928E-13m;
-        protected override decimal TebibytesPerSecondInOneBitPerSecond => 1.13686837721616E-13m;
+        protected override double TerabitsPerSecondInOneBitPerSecond => 1E-12;
+        protected override double TerabytesPerSecondInOneBitPerSecond => 1.25E-13;
+        protected override double TebibitsPerSecondInOneBitPerSecond => 9.09494701772928E-13;
+        protected override double TebibytesPerSecondInOneBitPerSecond => 1.13686837721616E-13;
 
-        protected override decimal PetabitsPerSecondInOneBitPerSecond => 1E-15m;
-        protected override decimal PetabytesPerSecondInOneBitPerSecond => 1.25E-16m;
-        protected override decimal PebibitsPerSecondInOneBitPerSecond => 8.88178419700125E-16m;
-        protected override decimal PebibytesPerSecondInOneBitPerSecond => 1.11022302462516E-16m;
+        protected override double PetabitsPerSecondInOneBitPerSecond => 1E-15;
+        protected override double PetabytesPerSecondInOneBitPerSecond => 1.25E-16;
+        protected override double PebibitsPerSecondInOneBitPerSecond => 8.88178419700125E-16;
+        protected override double PebibytesPerSecondInOneBitPerSecond => 1.11022302462516E-16;
 
-        protected override decimal ExabitsPerSecondInOneBitPerSecond => 1E-18m;
-        protected override decimal ExabytesPerSecondInOneBitPerSecond => 1.25E-19m;
-        protected override decimal ExbibitsPerSecondInOneBitPerSecond => 8.67361738E-19m;
-        protected override decimal ExbibytesPerSecondInOneBitPerSecond => 1.0842021724855E-19m;
+        protected override double ExabitsPerSecondInOneBitPerSecond => 1E-18;
+        protected override double ExabytesPerSecondInOneBitPerSecond => 1.25E-19;
+        protected override double ExbibitsPerSecondInOneBitPerSecond => 8.67361738E-19;
+        protected override double ExbibytesPerSecondInOneBitPerSecond => 1.0842021724855E-19;
     }
 }
